+++ pkg/xray/resource/resource_xray_webhook.go
@@ -108,7 +108,7 @@ func ResourceXrayWebhook() *schema.Resource {
 	var resourceXrayWebhookRead = func(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 		webhook := Webhook{}
 
-		req, err := getRestyRequest(m.(util.ProvderMetadata).Client, "")
+		req, err := getRestyRequest(m.(util.ProviderMetadata).Client, "")
 		if err != nil {
 			return diag.FromErr(err)
 		}
@@ -172,7 +172,7 @@ func ResourceXrayWebhook() *schema.Resource {
 			return diag.FromErr(err)
 		}
 
-		req, err := getRestyRequest(m.(util.ProvderMetadata).Client, "")
+		req, err := getRestyRequest(m.(util.ProviderMetadata).Client, "")
 		if err != nil {
 			return diag.FromErr(err)
 		}
@@ -198,7 +198,7 @@ func ResourceXrayWebhook() *schema.Resource {
 			return diag.FromErr(err)
 		}
 
-		req, err := getRestyRequest(m.(util.ProvderMetadata).Client, "")
+		req, err := getRestyRequest(m.(util.ProviderMetadata).Client, "")
 		if err != nil {
 			return diag.FromErr(err)
 		}
@@ -221,7 +221,7 @@ func ResourceXrayWebhook() *schema.Resource {
 	}
 
 	var resourceXrayWebhookDelete = func(_ context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
-		req, err := getRestyRequest(m.(util.ProvderMetadata).Client, "")
+		req, err := getRestyRequest(m.(util.ProviderMetadata).Client, "")
 		if err != nil {
 			return diag.FromErr(err)
 		}